requirements.txt
requests==2.24.0
Authlib==0.14.3
google-api-python-client==1.11.0
pandas==1.1.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
gspread~=3.6.0
pygsheets~=2.0.3.1